Add nonmember decorator to nested class in enum.
Why? The Info class isn't intended to be a member of the TpuChip enum.
In Python <=3.12, it actually is an enum member (though we don't care
that it is). It's just nested under that class to use the class as a
namespace. In Python 3.13, classes inside an enum will no longer be enum
members. This decorator just makes that explicit, so Python 3.12 doesn't
print warnings.