Only print core utilization for the first chip.
Why? Unfortunately, duty cycle is only reported per-chip. This means
that we're unable to see how busy an individual TPU core is. As a
workaround in tpu-info, we'll just report utilization per-chip.

In this commit, I also do some refactoring in order to accomodate
these changes. The refactoring wasn't strictly necessary for _just_ this
change, but I'm about to add some more functionality (i.e.
--output-format=json), which will benefit from this refactoring too.